requirements.txt
Babel==2.4.0
coverage==4.4.1
Django==1.11
django-address==0.1.5
django-betterforms==1.1.4
django-choices==1.6.0
django-countries==4.5
django-extensions==1.9.0
django-lockdown==1.4.2
django-phonenumber-field==1.3.0
django-postal==0.96
django-related-admin==0.4.0
graphviz==0.8
phonenumberslite==8.4.1
pydot==1.2.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
